requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
redis==3.5.3
celery==5.1.2
vk-api==11.9.9
python-telegram-bot==13.7
instagrapi==1.16.42
google-api-python-client==2.15.0
pinterest-api==0.0.8
alembic==1.7.1
pydantic==1.10.5
apscheduler==3.6.3
python-multipart==0.0.5
jinja2==3.0.1
python-dotenv==0.19.0
Pillow==9.5.0
pytz==2023.3